# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Work Agent",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/items": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Items",
        "operationId": "items_api_v1_items_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemView"
                  },
                  "title": "Response Items Api V1 Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{item_id}": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Get Item",
        "operationId": "get_item_api_v1_items__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "content"
        ],
        "summary": "Item Status",
        "operationId": "item_status_api_v1_items__item_id__patch",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{item_id}/file": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Get File",
        "operationId": "get_file_api_v1_items__item_id__file_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{item_id}/citation": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Citation",
        "operationId": "citation_api_v1_items__item_id__citation_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "locator",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Original",
              "title": "Locator"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitationView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{item_id}/analyze": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Analyze Item",
        "operationId": "analyze_item_api_v1_items__item_id__analyze_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Search",
        "operationId": "search_api_v1_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchHit"
                  },
                  "title": "Response Search Api V1 Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Citation": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "locator": {
            "type": "string",
            "title": "Locator",
            "default": "Original"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available"
          },
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          }
        },
        "type": "object",
        "required": [
          "item_id"
        ],
        "title": "Citation"
      },
      "CitationView": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "locator": {
            "type": "string",
            "title": "Locator"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "text",
          "locator",
          "version"
        ],
        "title": "CitationView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemMetadata": {
        "properties": {
          "demo": {
            "type": "boolean",
            "title": "Demo",
            "default": false
          },
          "start": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceTime"
              },
              {
                "type": "null"
              }
            ]
          },
          "end": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceTime"
              },
              {
                "type": "null"
              }
            ]
          },
          "due": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceTime"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ItemMetadata"
      },
      "ItemStatus": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ItemStatus"
      },
      "ItemView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "thread_key": {
            "type": "string",
            "title": "Thread Key"
          },
          "web_url": {
            "type": "string",
            "title": "Web Url"
          },
          "meta": {
            "$ref": "#/components/schemas/ItemMetadata"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "processing": {
            "type": "string",
            "title": "Processing"
          },
          "processing_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Error"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "external_id",
          "kind",
          "title",
          "body",
          "sender",
          "occurred_at",
          "thread_key",
          "web_url",
          "meta",
          "content_hash",
          "available",
          "status",
          "summary",
          "version",
          "processing",
          "source_name",
          "source_kind",
          "ai_enabled"
        ],
        "title": "ItemView"
      },
      "JobView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "dedupe_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dedupe Key"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "run_after": {
            "type": "string",
            "format": "date-time",
            "title": "Run After"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "payload",
          "status",
          "attempts",
          "run_after",
          "created_at"
        ],
        "title": "JobView"
      },
...
      "SearchHit": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "locator": {
            "type": "string",
            "title": "Locator"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "chunk_id",
          "title",
          "locator",
          "text",
          "version"
        ],
        "title": "SearchHit"
      },
...
      "SourceTime": {
        "properties": {
          "dateTime": {
            "type": "string",
            "title": "Datetime"
          },
          "timeZone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "type": "object",
        "required": [
          "dateTime"
        ],
        "title": "SourceTime"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}